2864 moving getToggleOptions out from BasicSchemaField
